Added solution to problem 37. Fixed solutions for problems 10 and 13.

+++ 1/Euler10.cpp
@@ -18,6 +18,6 @@
 
 #include "../helper.hpp"
 
-int solve10() {
+ull solve10() {
 	return sum( sieve( 2e6 ) );
 }
